Moved vim argument text object logic from `Buffer` to the vim crate
